+++ app/(2. pages)/(authentication)/page.tsx
@@ -2,7 +2,7 @@
 import { ReactNode } from "react";
 import { Center, Box } from "@mantine/core";
 import { loginLanguage } from "@shared/i18n/locales/fa";
-import LoginForm from "@widgets/login/login-form";
+import LoginForm from "@/3. widgets/login-form/ui";
 
 export default function Login(): ReactNode {
   return (
