+++ app/(2. pages)/panel/layout.tsx
@@ -1,8 +1,8 @@
 "use client";
 import React, { ReactNode } from "react";
 import { AppShell, Navbar, Header, MantineTheme, CSSObject } from "@mantine/core";
-import PanelNavbar from "@widgets/panel/panel-navbar";
-import PanelHeader from "@widgets/panel/panel-header";
+import PanelNavbar from "@/3. widgets/panel/ui/panel-navbar";
+import PanelHeader from "@/3. widgets/panel/ui/panel-header";
 
 const Layout = ({ children }: { children: ReactNode }): ReactNode => {
   return (
