+++ app/3. widgets/panel/ui/panel-header/index.tsx
